requirements.txt
django
django-bootstrap3
django-datetime-widget